Fermi/GBM: fix errors with response and background
The GBM response generation was incorrect because the wrong coordinate system for RA/Dec was being used (doh), and needlessly complicated.
The new version just uses the Sun center coordinate from `astropy.coordinates.get_sun`.
Further, the background subtraction was not ideal for this flare, and the event time didn not straddle a nonthermal peak.
So, those were also updated.
The fit results are significantly different.